feat(daemon): add single-instance coordination for tools
- Add ShowPreviewForPath daemon action for editor-to-daemon coordination
- Editor now opens preview via daemon to ensure single-instance behavior
- Close existing preview overlay before starting new capture
- Settings window uses single-instance check instead of NON_UNIQUE flag
- Editor uses NON_UNIQUE flag to allow multiple simultaneous editor instances
- Fix Qt6 type compatibility in C++ overlay (size() returns qsizetype)

+++ capture-overlay/src/CaptureOverlay_Drawing.cpp
@@ -2432,7 +2432,7 @@ void CaptureOverlay::drawKeystrokePreview(QPainter& p, double sx, double sy, dou
     QList<double> keyAlphas; // per-key fade alpha
     if (!m_keyPreviews.isEmpty()) {
         // Show live key presses (most recent up to 5)
-        int start = std::max(0, m_keyPreviews.size() - 5);
+        int start = std::max(0, static_cast<int>(m_keyPreviews.size()) - 5);
         for (int i = start; i < m_keyPreviews.size(); ++i) {
             displayKeys.append(m_keyPreviews[i].text);
             double age = (double)(now - m_keyPreviews[i].birthMs) / KEY_LIFETIME_MS;
